Introduced EditItem and made changes in logic.
+Added EditCartItem dialog.
+Added Discount TextField.
+Fixed problem with unsaved Product changes.
+Moved fragment of binding logic to fxml.

+++ src/main/java/pl/edu/agh/hibernate/example/shopapp/ShopDatabaseFiller.java
@@ -8,6 +8,7 @@ import pl.edu.agh.hibernate.example.shopapp.model.product.Product;
 
 import javax.persistence.EntityManager;
 import javax.persistence.EntityTransaction;
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -36,8 +37,8 @@ public class ShopDatabaseFiller {
 
     private static List<Customer> createCustomers() {
         List<Customer> customers = new ArrayList<>();
-        customers.add(new Customer("Joseph Black", new Address("Soft River Str.", "Zywiec", "34-300"), 0.05));
-        customers.add(new Customer("Pawel Pablo", new Address("Blue Str.", "Bielsko-Biala", "33-300"), 0.1));
+        customers.add(new Customer("Joseph Black", new Address("Soft River Str.", "Zywiec", "34-300"), BigDecimal.valueOf(0.05)));
+        customers.add(new Customer("Pawel Pablo", new Address("Blue Str.", "Bielsko-Biala", "33-300"), BigDecimal.valueOf(0.1)));
         return customers;
     }
 
@@ -51,11 +52,11 @@ public class ShopDatabaseFiller {
 
     private static List<Product> createProducts(List<Supplier> suppliers, List<Category> categories) {
         List<Product> products = new ArrayList<>();
-        products.add(new Product("Teddy Bear", categories.get(0), 10, suppliers.get(0), 10.0));
-        products.add(new Product("Doll", categories.get(0), 1, suppliers.get(0), 24.0));
-        products.add(new Product("Apple", categories.get(1), 100, suppliers.get(1), 2.3));
-        products.add(new Product("Peach", categories.get(1), 23, suppliers.get(1), 3.68));
-        products.add(new Product("Carrot", categories.get(2), 3, suppliers.get(1), 0.89));
+        products.add(new Product("Teddy Bear", categories.get(0), 10, suppliers.get(0), BigDecimal.valueOf(10.0)));
+        products.add(new Product("Doll", categories.get(0), 1, suppliers.get(0), BigDecimal.valueOf(24.0)));
+        products.add(new Product("Apple", categories.get(1), 100, suppliers.get(1), BigDecimal.valueOf(2.3)));
+        products.add(new Product("Peach", categories.get(1), 23, suppliers.get(1), BigDecimal.valueOf(3.68)));
+        products.add(new Product("Carrot", categories.get(2), 3, suppliers.get(1), BigDecimal.valueOf(0.89)));
         return products;
     }
